ml-engine/requirements.txt
# Peterbilt Sentiment Analyzer - ML Engine Dependencies

# Web framework
flask>=3.0.0
gunicorn>=21.0.0

# ML/NLP
scikit-learn>=1.4.0
nltk>=3.8.0
pandas>=2.1.0
gunicorn>=21.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0

# Environment
python-dotenv>=1.0.0

# Database (Supabase)
supabase>=2.0.0

# Testing
pytest>=7.0.0
